fix: convert @apilink to @link on build
Related: https://github.com/apify/crawlee/issues/2717
