+++ package.json
@@ -82,6 +82,7 @@
         "eslint-config-prettier": "^10.1.1",
         "fs-extra": "^11.2.0",
         "gen-esm-wrapper": "^1.1.3",
+        "globby": "^14.1.0",
         "husky": "^9.1.7",
         "lerna": "^8.1.8",
         "lint-staged": "^15.2.8",
