--- scripts/temp_fix_apify_exports.ts
@@ -1,14 +1,0 @@
-import { copyFileSync, readFileSync, writeFileSync } from 'fs';
-import { resolve } from 'path';
-
-const target = resolve(process.cwd(), 'dist', 'index.d.ts');
-
-const file = readFileSync(target).toString();
-
-writeFileSync(
-    target,
-    file.replace(
-        `export * from './exports';`,
-        `// @ts-ignore\nexport * from './exports';`,
-    ),
-);
